feat: added possibility to copy files from several folders

+++ SimpleArchiver/Program.cs
@@ -1,24 +1,49 @@
-﻿// See https://aka.ms/new-console-template for more information
-
-using Newtonsoft.Json;
+﻿using Newtonsoft.Json;
 using SimpleArchiver.Models;
 
 var config = GetConfig();
-var sourceFiles = Directory.GetFiles(config.SourceFolder);
+var sourceFiles = GetFiles(config.SourceFolders);
 
-//folder name can't contain ":" 
+//folder name can't contain ":"
 var targetFolderTimeStamp = config.TargetFolder + $"\\{DateTime.Now:G}".Replace(":", "_");
 Directory.CreateDirectory(targetFolderTimeStamp);
+
 foreach (var sourceFile in sourceFiles)
 {
-    var targetFileName = sourceFile.Replace(config.SourceFolder, targetFolderTimeStamp);
-    File.Copy(sourceFile, targetFileName);
+    var sourceFolder = Path.GetDirectoryName(sourceFile);
+    var targetFileName = sourceFile.Replace(sourceFolder, targetFolderTimeStamp);
+    if (!File.Exists(targetFileName))
+    {
+        File.Copy(sourceFile, targetFileName);
+    }
 }
 
 Config GetConfig()
 {
+    //this config is located in same directory with SimpleArchiver.exe
     var jsonConfig = File.ReadAllText("config.json");
     var configModel = JsonConvert.DeserializeObject<Config>(jsonConfig);
 
     return configModel;
+}
+
+List<string> GetFiles(List<string> paths)
+{
+    var resultFiles = new List<string>();
+    foreach (var path in paths)
+    {
+        try
+        {
+            var currentFiles = Directory.GetFiles(path);
+            resultFiles.AddRange(currentFiles);
+        }
+        catch (IOException e)
+        {
+            Console.WriteLine(e.Message);
+            Console.ReadLine();
+            throw;
+        }
+    }
+
+    return resultFiles;
 }
